Exercise 4 Completed
The Exercise 4 was completed. The Oled is being refreshed using the data
from the SRAM. There was a wiring mistake which was corrected. The
refreshing is faster than the one we used to do by creating on the
moment the menu and then sending it to the Oled. The little flickering
the Oled had was corrected with this new version